requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
numbers-parser>=4.16.0
openpyxl>=3.1.0
PyGithub>=2.0.0
